fix(geofence): stamp the foreground re-add for the contradiction gate
The foreground rearm is a third (re)add site, and a re-add can replay
the daemon's wedged belief before its fresh evaluation — the gate must
vet those events against this add's window and geometry like the other
two sites. Splits the monitor's fix-reading trio into +Fixes.swift to
stay under the file-length cap now that both features share the file.
Both changes ran all weekend on the combined field build.
